chore: refactor pool logic
