+++ cmd/loadtest/uniswapv3.go
@@ -18,6 +18,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
+	uniswapv3loadtest "github.com/maticnetwork/polygon-cli/cmd/loadtest/uniswapv3"
 	"github.com/maticnetwork/polygon-cli/contracts/uniswapv3"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
@@ -46,13 +47,6 @@ const (
 	// The max duration of an incentive in seconds.
 	// https://github.com/Uniswap/deploy-v3/blob/b7aac0f1c5353b36802dc0cf95c426d2ef0c3252/src/steps/deploy-v3-staker.ts#L13
 	MAX_INCENTIVE_DURATION = ONE_YEAR_SECONDS * 2
-
-	// The minimum tick that may be passed to `getSqrtRatioAtTick` computed from log base 1.0001 of 2**-128.
-	// https://github.com/Uniswap/v3-core/blob/d8b1c635c275d2a9450bd6a78f3fa2484fef73eb/contracts/libraries/TickMath.sol#L9
-	MIN_TICK = -887272
-	// The maximum tick that may be passed to `getSqrtRatioAtTick` computed from log base 1.0001 of 2**128.
-	// https://github.com/Uniswap/v3-core/blob/d8b1c635c275d2a9450bd6a78f3fa2484fef73eb/contracts/libraries/TickMath.sol#L11
-	MAX_TICK = -MIN_TICK
 )
 
 var (
@@ -143,79 +137,14 @@ func init() {
 	uniswapv3LoadTestParams = *params
 }
 
-type UniswapV3Addresses struct {
-	FactoryV3, Multicall, ProxyAdmin, TickLens, NFTDescriptorLib, NonfungibleTokenPositionDescriptor, TransparentUpgradeableProxy, NonfungiblePositionManager, Migrator, Staker, QuoterV2, SwapRouter02 common.Address
-	WETH9                                                                                                                                                                                               common.Address
-}
-
-type UniswapV3Config struct {
-	FactoryV3                          contractConfig[uniswapv3.UniswapV3Factory]
-	Multicall                          contractConfig[uniswapv3.UniswapInterfaceMulticall]
-	ProxyAdmin                         contractConfig[uniswapv3.ProxyAdmin]
-	TickLens                           contractConfig[uniswapv3.TickLens]
-	NFTDescriptorLib                   contractConfig[uniswapv3.NFTDescriptor]
-	NonfungibleTokenPositionDescriptor contractConfig[uniswapv3.NonfungibleTokenPositionDescriptor]
-	TransparentUpgradeableProxy        contractConfig[uniswapv3.TransparentUpgradeableProxy]
-	NonfungiblePositionManager         contractConfig[uniswapv3.NonfungiblePositionManager]
-	Migrator                           contractConfig[uniswapv3.V3Migrator]
-	Staker                             contractConfig[uniswapv3.UniswapV3Staker]
-	QuoterV2                           contractConfig[uniswapv3.QuoterV2]
-	SwapRouter02                       contractConfig[uniswapv3.SwapRouter02]
-
-	WETH9 contractConfig[uniswapv3.WETH9]
-}
-
-func (c *UniswapV3Config) ToAddresses() UniswapV3Addresses {
-	return UniswapV3Addresses{
-		FactoryV3:                          c.FactoryV3.Address,
-		Multicall:                          c.Multicall.Address,
-		ProxyAdmin:                         c.ProxyAdmin.Address,
-		TickLens:                           c.TickLens.Address,
-		NFTDescriptorLib:                   c.NFTDescriptorLib.Address,
-		NonfungibleTokenPositionDescriptor: c.NonfungibleTokenPositionDescriptor.Address,
-		TransparentUpgradeableProxy:        c.TransparentUpgradeableProxy.Address,
-		NonfungiblePositionManager:         c.NonfungiblePositionManager.Address,
-		Migrator:                           c.Migrator.Address,
-		Staker:                             c.Staker.Address,
-		QuoterV2:                           c.QuoterV2.Address,
-		SwapRouter02:                       c.SwapRouter02.Address,
-		WETH9:                              c.WETH9.Address,
-	}
-}
-
-type PoolConfig struct {
-	Token0, Token1     contractConfig[uniswapv3.Swapper]
-	ReserveA, ReserveB *big.Int
-	Fees               *big.Int
-}
-
-type contractConfig[T uniswapV3Contract] struct {
-	Address  common.Address
-	contract *T
-}
-
-type uniswapV3Contract interface {
-	uniswapv3.UniswapV3Factory | uniswapv3.UniswapInterfaceMulticall | uniswapv3.ProxyAdmin | uniswapv3.TickLens | uniswapv3.WETH9 | uniswapv3.NFTDescriptor | uniswapv3.NonfungibleTokenPositionDescriptor | uniswapv3.TransparentUpgradeableProxy | uniswapv3.NonfungiblePositionManager | uniswapv3.V3Migrator | uniswapv3.UniswapV3Staker | uniswapv3.QuoterV2 | uniswapv3.SwapRouter02 | uniswapv3.Swapper
-}
-
-type slot struct {
-	SqrtPriceX96               *big.Int
-	Tick                       *big.Int
-	ObservationIndex           uint16
-	ObservationCardinality     uint16
-	ObservationCardinalityNext uint16
-	FeeProtocol                uint8
-	Unlocked                   bool
-}
-
 // Deploy the full UniswapV3 contract suite in 15 different steps.
 // Source: https://github.com/Uniswap/deploy-v3
-func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, knownAddresses UniswapV3Addresses, ownerAddress common.Address) (UniswapV3Config, error) {
-	config := UniswapV3Config{}
+func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, knownAddresses uniswapv3loadtest.Address, ownerAddress common.Address) (uniswapv3loadtest.UniswapV3Config, error) {
+	config := uniswapv3loadtest.UniswapV3Config{}
 	var err error
 
 	// 0. Deploy WETH9.
-	config.WETH9.Address, config.WETH9.contract, err = deployOrInstantiateContract(
+	config.WETH9.Address, config.WETH9.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 0: Contract WETH9 deployment", knownAddresses.WETH9,
 		uniswapv3.DeployWETH9,
 		uniswapv3.NewWETH9,
@@ -229,7 +158,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 1. Deploy UniswapV3Factory.
-	config.FactoryV3.Address, config.FactoryV3.contract, err = deployOrInstantiateContract(
+	config.FactoryV3.Address, config.FactoryV3.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 1: Contract UniswapV3Factory deployment", knownAddresses.FactoryV3,
 		uniswapv3.DeployUniswapV3Factory,
 		uniswapv3.NewUniswapV3Factory,
@@ -243,12 +172,12 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 2. Enable one basic point fee tier.
-	if err = enableOneBPFeeTier(config.FactoryV3.contract, tops, cops, ONE_BP_FEE, ONE_BP_TICK_SPACING); err != nil {
+	if err = enableOneBPFeeTier(config.FactoryV3.Contract, tops, cops, ONE_BP_FEE, ONE_BP_TICK_SPACING); err != nil {
 		return config, err
 	}
 
 	// 3. Deploy UniswapInterfaceMulticall.
-	config.Multicall.Address, config.Multicall.contract, err = deployOrInstantiateContract(
+	config.Multicall.Address, config.Multicall.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 3: Contract UniswapInterfaceMulticall deployment", knownAddresses.Multicall,
 		uniswapv3.DeployUniswapInterfaceMulticall,
 		uniswapv3.NewUniswapInterfaceMulticall,
@@ -262,7 +191,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 4. Deploy ProxyAdmin.
-	config.ProxyAdmin.Address, config.ProxyAdmin.contract, err = deployOrInstantiateContract(
+	config.ProxyAdmin.Address, config.ProxyAdmin.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 4: Contract ProxyAdmin deployment", knownAddresses.ProxyAdmin,
 		uniswapv3.DeployProxyAdmin,
 		uniswapv3.NewProxyAdmin,
@@ -276,7 +205,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 5. Deploy TickLens.
-	config.TickLens.Address, config.TickLens.contract, err = deployOrInstantiateContract(
+	config.TickLens.Address, config.TickLens.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 5: Contract TickLens deployment", knownAddresses.TickLens,
 		uniswapv3.DeployTickLens,
 		uniswapv3.NewTickLens,
@@ -308,7 +237,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 7. Deploy NonfungibleTokenPositionDescriptor.
-	config.NonfungibleTokenPositionDescriptor.Address, config.NonfungibleTokenPositionDescriptor.contract, err = deployOrInstantiateContract(
+	config.NonfungibleTokenPositionDescriptor.Address, config.NonfungibleTokenPositionDescriptor.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 7: Contract NonfungibleTokenPositionDescriptor deployment", knownAddresses.NonfungibleTokenPositionDescriptor,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.NonfungibleTokenPositionDescriptor, error) {
 			oldAddressFmt := "__$cea9be979eee3d87fb124d6cbb244bb0b5$__"
@@ -339,7 +268,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 8. Deploy TransparentUpgradeableProxy.
-	config.TransparentUpgradeableProxy.Address, config.TransparentUpgradeableProxy.contract, err = deployOrInstantiateContract(
+	config.TransparentUpgradeableProxy.Address, config.TransparentUpgradeableProxy.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 8: Contract TransparentUpgradeableProxy deployment", knownAddresses.TransparentUpgradeableProxy,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.TransparentUpgradeableProxy, error) {
 			return uniswapv3.DeployTransparentUpgradeableProxy(tops, c, config.NonfungibleTokenPositionDescriptor.Address, config.ProxyAdmin.Address, []byte(""))
@@ -363,7 +292,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 9. Deploy NonfungiblePositionManager.
-	config.NonfungiblePositionManager.Address, config.NonfungiblePositionManager.contract, err = deployOrInstantiateContract(
+	config.NonfungiblePositionManager.Address, config.NonfungiblePositionManager.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 9: Contract NonfungiblePositionManager deployment", knownAddresses.NonfungiblePositionManager,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.NonfungiblePositionManager, error) {
 			return uniswapv3.DeployNonfungiblePositionManager(tops, c, config.FactoryV3.Address, config.WETH9.Address, config.TransparentUpgradeableProxy.Address)
@@ -379,7 +308,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 10. Deploy Migrator.
-	config.Migrator.Address, config.Migrator.contract, err = deployOrInstantiateContract(
+	config.Migrator.Address, config.Migrator.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 10: Contract V3Migrator deployment", knownAddresses.Migrator,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.V3Migrator, error) {
 			return uniswapv3.DeployV3Migrator(tops, c, config.FactoryV3.Address, config.WETH9.Address, config.NonfungiblePositionManager.Address)
@@ -395,12 +324,12 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 11. Set Factory owner.
-	if err = setFactoryOwner(config.FactoryV3.contract, tops, cops, ownerAddress); err != nil {
+	if err = setFactoryOwner(config.FactoryV3.Contract, tops, cops, ownerAddress); err != nil {
 		return config, err
 	}
 
 	// 12. Deploy Staker.
-	config.Staker.Address, config.Staker.contract, err = deployOrInstantiateContract(
+	config.Staker.Address, config.Staker.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 12: Contract UniswapV3Staker deployment", knownAddresses.Staker,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.UniswapV3Staker, error) {
 			return uniswapv3.DeployUniswapV3Staker(tops, c, config.FactoryV3.Address, config.NonfungiblePositionManager.Address, big.NewInt(MAX_INCENTIVE_START_LEAD_TIME), big.NewInt(MAX_INCENTIVE_DURATION))
@@ -416,7 +345,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 13. Deploy QuoterV2.
-	config.QuoterV2.Address, config.QuoterV2.contract, err = deployOrInstantiateContract(
+	config.QuoterV2.Address, config.QuoterV2.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 13: Contract QuoterV2 deployment", knownAddresses.QuoterV2,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.QuoterV2, error) {
 			return uniswapv3.DeployQuoterV2(tops, c, config.FactoryV3.Address, config.WETH9.Address)
@@ -432,7 +361,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 14. Deploy SwapRouter02.
-	config.SwapRouter02.Address, config.SwapRouter02.contract, err = deployOrInstantiateContract(
+	config.SwapRouter02.Address, config.SwapRouter02.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, "Step 14: Contract SwapRouter02 deployment", knownAddresses.SwapRouter02,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.SwapRouter02, error) {
 			// Note: we specify an empty address for UniswapV2Factory.
@@ -450,7 +379,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 	}
 
 	// 15. Transfer ProxyAdmin ownership.
-	if err = transferProxyAdminOwnership(config.ProxyAdmin.contract, tops, cops, ownerAddress); err != nil {
+	if err = transferProxyAdminOwnership(config.ProxyAdmin.Contract, tops, cops, ownerAddress); err != nil {
 		return config, err
 	}
 
@@ -460,7 +389,7 @@ func deployUniswapV3(ctx context.Context, c *ethclient.Client, tops *bind.Transa
 // Deploy or instantiate any UniswapV3 contract.
 // This method will either deploy a contract if the known address is empty (equal to `common.Address{}` or `0x0“)
 // or instantiate the contract if the known address is specified.
-func deployOrInstantiateContract[T uniswapV3Contract](
+func deployOrInstantiateContract[T uniswapv3loadtest.Contract](
 	ctx context.Context,
 	c *ethclient.Client,
 	tops *bind.TransactOpts,
@@ -566,11 +495,11 @@ func transferProxyAdminOwnership(contract *uniswapv3.ProxyAdmin, tops *bind.Tran
 	return nil
 }
 
-func deploySwapperContract(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, config UniswapV3Config, tokenName, tokenSymbol string, recipient common.Address, tokenKnownAddress common.Address) (contractConfig[uniswapv3.Swapper], error) {
-	var token contractConfig[uniswapv3.Swapper]
+func deploySwapperContract(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, config uniswapv3loadtest.UniswapV3Config, tokenName, tokenSymbol string, recipient common.Address, tokenKnownAddress common.Address) (uniswapv3loadtest.ContractConfig[uniswapv3.Swapper], error) {
+	var token uniswapv3loadtest.ContractConfig[uniswapv3.Swapper]
 	var err error
 	addressesToApprove := []common.Address{config.NonfungiblePositionManager.Address, config.SwapRouter02.Address}
-	token.Address, token.contract, err = deployOrInstantiateContract(
+	token.Address, token.Contract, err = deployOrInstantiateContract(
 		ctx, c, tops, cops, tokenName, tokenKnownAddress,
 		func(*bind.TransactOpts, bind.ContractBackend) (common.Address, *types.Transaction, *uniswapv3.Swapper, error) {
 			return uniswapv3.DeploySwapper(tops, c)
@@ -617,151 +546,8 @@ func approveSwapperSpendingsByUniswap(ctx context.Context, contract *uniswapv3.S
 	return nil
 }
 
-// createPool sets up a Uniswap V3 pool by creating it, initializing it if necessary and providing
-// liquidity if none exists.
-func createPool(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, uniswapV3Config UniswapV3Config, poolConfig PoolConfig, recipient common.Address) error {
-	// Create and initialize the pool.
-	// No need to check if the pool was already created or initialized, the contract handles every scenario.
-	sqrtPriceX96 := computeSqrtPriceX96(poolConfig.ReserveA, poolConfig.ReserveB)
-	if _, err := uniswapV3Config.NonfungiblePositionManager.contract.CreateAndInitializePoolIfNecessary(tops, poolConfig.Token0.Address, poolConfig.Token1.Address, poolConfig.Fees, sqrtPriceX96); err != nil {
-		log.Error().Err(err).Msg("Unable to create and initialize the pool")
-		return err
-	}
-	log.Debug().Msg("Pool created and initialized (if necessary)")
-
-	// Retrieve the pool address.
-	var poolAddress common.Address
-	if err := blockUntilSuccessful(ctx, c, func() (err error) {
-		poolAddress, err = uniswapV3Config.FactoryV3.contract.GetPool(cops, poolConfig.Token0.Address, poolConfig.Token1.Address, poolConfig.Fees)
-		if poolAddress == (common.Address{}) {
-			return fmt.Errorf("pool not deployed yet")
-		}
-		return
-	}); err != nil {
-		log.Error().Err(err).Msg("Unable to retrieve the address of the pool")
-		return err
-	}
-
-	// Instantiate the pool contract.
-	poolContract, err := uniswapv3.NewIUniswapV3Pool(poolAddress, c)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to instantiate the pool")
-		return err
-	}
-	log.Debug().Interface("address", poolAddress).Msg("Pool instantiated")
-
-	// Get pool state.
-	var slot0 slot
-	slot0, err = poolContract.Slot0(cops)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to get pool's slot0")
-		return err
-	}
-
-	// Check pool's liquidity.
-	var liquidity *big.Int
-	liquidity, err = poolContract.Liquidity(cops)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to get pool's liquidity")
-		return err
-	}
-	log.Debug().Interface("slot0", slot0).Interface("liquidity", liquidity).Msg("Pool state")
-
-	// Provide liquidity if there's none.
-	if liquidity.Cmp(big.NewInt(0)) == 0 {
-		// Compute the tick lower and upper for providing liquidity.
-		// The default tick spacing is set to 60 for the 0.3% fee tier and unfortunately, `MIN_TICK` and
-		// `MAX_TICK` are not divisible by this amount. The solution is to use a multiple of 60 instead.
-		var tickSpacing *big.Int
-		tickSpacing, err = poolContract.TickSpacing(cops)
-		if err != nil {
-			log.Error().Err(err).Msg("Unable to get tick spacing")
-			return err
-		}
-		// tickUpper = (MAX_TICK / tickSpacing) * tickSpacing
-		tickUpper := new(big.Int).Div(big.NewInt(MAX_TICK), tickSpacing)
-		tickUpper.Mul(tickUpper, tickSpacing)
-		// tickLower = - tickUpper
-		tickLower := new(big.Int).Neg(tickUpper)
-
-		// Provide liquidity.
-		dl, _ := big.NewInt(0).SetString("115792089237316195423570985008687907853269984665640564039457584007913129639935", 10)
-		amMax := new(big.Int).Div(tokenPoolSize, big.NewInt(2))
-		amMin, _ := big.NewInt(0).SetString("1", 10)
-		mintParams := uniswapv3.INonfungiblePositionManagerMintParams{
-			// The address of the token0, first token of the pool.
-			Token0: poolConfig.Token0.Address,
-			// The address of the token1, second token of the pool.
-			Token1: poolConfig.Token1.Address,
-			// The fee associated with the pool.
-			Fee: poolConfig.Fees,
-
-			// The lower end of the tick range for the position.
-			// Here, we provide liquidity across the whole possible range (divisible by tick spacing).
-			TickLower: tickLower,
-			// The higher end of the tick range for the position.
-			TickUpper: tickUpper,
-
-			// The desired amount of token0 to be sent to the pool during the minting operation.
-			Amount0Desired: amMax,
-			// The desired amount of token1 to be sent to the pool during the minting operation.
-			Amount1Desired: amMax,
-			// The minimum acceptable amount of token0 to be sent to the pool. This represents the slippage
-			// protection for token0 during the minting. Here, we mint without any slippage protection.
-			// Don't do this in production!
-			Amount0Min: amMin,
-			// The minimum acceptable amount of token1 to be sent to the pool.
-			Amount1Min: amMin,
-
-			// The destination address of the pool fees.
-			Recipient: recipient,
-
-			// The unix time after which the mint will fail, to protect against long-pending transactions
-			// and wild swings in prices.
-			Deadline: dl, // 10 minutes to execute the swap.
-			// Deadline: big.NewInt(1759474606), // in 2 years (2025-10-03)
-		}
-		if err := blockUntilSuccessful(ctx, c, func() (err error) {
-			// Mint tokens.
-			_, err = uniswapV3Config.NonfungiblePositionManager.contract.Mint(tops, mintParams)
-			if err != nil {
-				return err
-			}
-
-			// Check that liquidity has been added to the pool.
-			liquidity, err = poolContract.Liquidity(cops)
-			if err != nil {
-				return err
-			}
-			if liquidity.Cmp(big.NewInt(0)) == 0 {
-				return errors.New("pool has no liquidity")
-			}
-			return nil
-		}); err != nil {
-			log.Error().Err(err).Msg("Unable to provide liquidity to the pool")
-			return err
-		}
-		log.Debug().Interface("liquidity", liquidity).Msg("Liquidity provided to the pool")
-	} else {
-		log.Debug().Msg("Liquidity already provided to the pool")
-	}
-
-	return nil
-}
-
-// computeSqrtPriceX96 calcules the square root of the price ratio of two reserves in a UniswapV3 pool.
-// https://uniswapv3book.com/docs/milestone_1/calculating-liquidity/#price-range-calculation
-func computeSqrtPriceX96(reserveA, reserveB *big.Int) *big.Int {
-	sqrtReserveA := new(big.Int).Sqrt(reserveA)
-	sqrtReserveB := new(big.Int).Sqrt(reserveB)
-	q96 := new(big.Int).Exp(big.NewInt(2), big.NewInt(96), nil)
-	sqrtPriceX96 := new(big.Int).Mul(sqrtReserveB, q96)
-	sqrtPriceX96.Div(sqrtPriceX96, sqrtReserveA)
-	return sqrtPriceX96
-}
-
 // Run UniswapV3 loadtest by performing swaps.
-func loadTestUniswapV3(ctx context.Context, c *ethclient.Client, nonce uint64, uniswapV3Config UniswapV3Config, poolConfig PoolConfig) (t1 time.Time, t2 time.Time, err error) {
+func loadTestUniswapV3(ctx context.Context, c *ethclient.Client, nonce uint64, uniswapV3Config uniswapv3loadtest.UniswapV3Config, poolConfig uniswapv3loadtest.PoolConfig) (t1 time.Time, t2 time.Time, err error) {
 	ltp := inputLoadTestParams
 	chainID := new(big.Int).SetUint64(*ltp.ChainID)
 	privateKey := ltp.ECDSAPrivateKey
@@ -776,14 +562,14 @@ func loadTestUniswapV3(ctx context.Context, c *ethclient.Client, nonce uint64, u
 
 	t1 = time.Now()
 	defer func() { t2 = time.Now() }()
-	err = exactInputSingleSwap(tops, uniswapV3Config.SwapRouter02.contract, poolConfig, *ltp.FromETHAddress, nonce)
+	err = exactInputSingleSwap(tops, uniswapV3Config.SwapRouter02.Contract, poolConfig, *ltp.FromETHAddress, nonce)
 	return
 }
 
 // exactInputSingleSwap performs a UniswapV3 swap using the `ExactInputSingle` method which swaps a fixed amount of
 // one token for a maximum possible amount of another token. The direction of the swap is determined
 // by the nonce value.
-func exactInputSingleSwap(tops *bind.TransactOpts, swapRouter *uniswapv3.SwapRouter02, poolConfig PoolConfig, recipient common.Address, nonce uint64) error {
+func exactInputSingleSwap(tops *bind.TransactOpts, swapRouter *uniswapv3.SwapRouter02, poolConfig uniswapv3loadtest.PoolConfig, recipient common.Address, nonce uint64) error {
 	// Determine the direction of the swap.
 	tIn := poolConfig.Token0
 	tInName := "token0"
