+++ cmd/loadtest/uniswapv3/pool.go
@@ -1,0 +1,212 @@
+package uniswapv3loadtest
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/maticnetwork/polygon-cli/contracts/uniswapv3"
+	"github.com/rs/zerolog/log"
+)
+
+const (
+	// The minimum tick that may be passed to `getSqrtRatioAtTick` computed from log base 1.0001 of 2**-128.
+	// https://github.com/Uniswap/v3-core/blob/d8b1c635c275d2a9450bd6a78f3fa2484fef73eb/contracts/libraries/TickMath.sol#L9
+	MIN_TICK = -887272
+	// The maximum tick that may be passed to `getSqrtRatioAtTick` computed from log base 1.0001 of 2**128.
+	// https://github.com/Uniswap/v3-core/blob/d8b1c635c275d2a9450bd6a78f3fa2484fef73eb/contracts/libraries/TickMath.sol#L11
+	MAX_TICK = -MIN_TICK
+)
+
+// PoolConfig represents the configuration of a UniswapV3 pool.
+type PoolConfig struct {
+	Token0, Token1     ContractConfig[uniswapv3.Swapper]
+	ReserveA, ReserveB *big.Int
+	Fees               *big.Int
+}
+
+// slot represents the state of a UniswapV3 pool.
+type slot struct {
+	SqrtPriceX96               *big.Int
+	Tick                       *big.Int
+	ObservationIndex           uint16
+	ObservationCardinality     uint16
+	ObservationCardinalityNext uint16
+	FeeProtocol                uint8
+	Unlocked                   bool
+}
+
+type blockUntilSuccessfulFn func(ctx context.Context, c *ethclient.Client, f func() error) error
+
+// CreatePool sets up a Uniswap V3 pool by creating it, initializing it if necessary and providing
+// liquidity if none exists.
+func SetupPool(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, uniswapV3Config UniswapV3Config, poolConfig PoolConfig, size *big.Int, recipient common.Address, blockUntilSuccessful blockUntilSuccessfulFn) error {
+	poolContract, err := createPool(ctx, c, tops, cops, uniswapV3Config, poolConfig, blockUntilSuccessful)
+	if err != nil {
+		return err
+	}
+
+	var slot0 slot
+	var liquidity *big.Int
+	slot0, liquidity, err = getPoolState(cops, poolContract)
+	if err != nil {
+		return err
+	}
+	log.Debug().Interface("slot0", slot0).Interface("liquidity", liquidity).Msg("Pool state")
+
+	// Provide liquidity if there's none.
+	if liquidity.Cmp(big.NewInt(0)) == 0 {
+		provideLiquidity(ctx, c, tops, cops, poolContract, poolConfig, size, recipient, uniswapV3Config.NonfungiblePositionManager.Contract, blockUntilSuccessful)
+	} else {
+		log.Debug().Msg("Liquidity already provided to the pool")
+	}
+
+	return nil
+}
+
+// createPool creates and initialises the pool if necessary.
+func createPool(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, uniswapV3Config UniswapV3Config, poolConfig PoolConfig, blockUntilSuccessful blockUntilSuccessfulFn) (*uniswapv3.IUniswapV3Pool, error) {
+	// Create and initialize the pool.
+	sqrtPriceX96 := computeSqrtPriceX96(poolConfig.ReserveA, poolConfig.ReserveB)
+	if _, err := uniswapV3Config.NonfungiblePositionManager.Contract.CreateAndInitializePoolIfNecessary(tops, poolConfig.Token0.Address, poolConfig.Token1.Address, poolConfig.Fees, sqrtPriceX96); err != nil {
+		log.Error().Err(err).Msg("Unable to create and initialize the pool")
+		return nil, err
+	}
+	log.Debug().Msg("Pool created and initialized (if necessary)")
+
+	// Retrieve the pool address.
+	var poolAddress common.Address
+	if err := blockUntilSuccessful(ctx, c, func() (err error) {
+		poolAddress, err = uniswapV3Config.FactoryV3.Contract.GetPool(cops, poolConfig.Token0.Address, poolConfig.Token1.Address, poolConfig.Fees)
+		if poolAddress == (common.Address{}) {
+			return fmt.Errorf("pool not deployed yet")
+		}
+		return
+	}); err != nil {
+		log.Error().Err(err).Msg("Unable to retrieve the address of the pool")
+		return nil, err
+	}
+
+	// Instantiate the pool contract.
+	contract, err := uniswapv3.NewIUniswapV3Pool(poolAddress, c)
+	if err != nil {
+		log.Error().Err(err).Msg("Unable to instantiate the pool")
+		return nil, err
+	}
+	log.Debug().Interface("address", poolAddress).Msg("Pool instantiated")
+	return contract, nil
+}
+
+// computeSqrtPriceX96 calcules the square root of the price ratio of two reserves in a UniswapV3 pool.
+// https://uniswapv3book.com/docs/milestone_1/calculating-liquidity/#price-range-calculation
+func computeSqrtPriceX96(reserveA, reserveB *big.Int) *big.Int {
+	sqrtReserveA := new(big.Int).Sqrt(reserveA)
+	sqrtReserveB := new(big.Int).Sqrt(reserveB)
+	q96 := new(big.Int).Exp(big.NewInt(2), big.NewInt(96), nil)
+	sqrtPriceX96 := new(big.Int).Mul(sqrtReserveB, q96)
+	sqrtPriceX96.Div(sqrtPriceX96, sqrtReserveA)
+	return sqrtPriceX96
+}
+
+// getPoolState returns pool's slot0 and liquidity.
+func getPoolState(cops *bind.CallOpts, contract *uniswapv3.IUniswapV3Pool) (slot, *big.Int, error) {
+	// Get pool state.
+	var slot0 slot
+	var err error
+	slot0, err = contract.Slot0(cops)
+	if err != nil {
+		log.Error().Err(err).Msg("Unable to get pool's slot0")
+		return slot{}, nil, err
+	}
+
+	// Get pool's liquidity.
+	var liquidity *big.Int
+	liquidity, err = contract.Liquidity(cops)
+	if err != nil {
+		log.Error().Err(err).Msg("Unable to get pool's liquidity")
+		return slot{}, nil, err
+	}
+
+	return slot0, liquidity, nil
+}
+
+// provideLiquidity provides liquidity to the pool.
+func provideLiquidity(ctx context.Context, c *ethclient.Client, tops *bind.TransactOpts, cops *bind.CallOpts, poolContract *uniswapv3.IUniswapV3Pool, poolConfig PoolConfig, poolSize *big.Int, recipient common.Address, nftPositionManagerContract *uniswapv3.NonfungiblePositionManager, blockUntilSuccessful blockUntilSuccessfulFn) error {
+	// Compute the tick lower and upper for providing liquidity.
+	// The default tick spacing is set to 60 for the 0.3% fee tier and unfortunately, `MIN_TICK` and
+	// `MAX_TICK` are not divisible by this amount. The solution is to use a multiple of 60 instead.
+	tickSpacing, err := poolContract.TickSpacing(cops)
+	if err != nil {
+		log.Error().Err(err).Msg("Unable to get tick spacing")
+		return err
+	}
+	// tickUpper = (MAX_TICK / tickSpacing) * tickSpacing
+	tickUpper := new(big.Int).Div(big.NewInt(MAX_TICK), tickSpacing)
+	tickUpper.Mul(tickUpper, tickSpacing)
+	// tickLower = - tickUpper
+	tickLower := new(big.Int).Neg(tickUpper)
+
+	// Provide liquidity.
+	var liquidity *big.Int
+	dl, _ := big.NewInt(0).SetString("115792089237316195423570985008687907853269984665640564039457584007913129639935", 10)
+	amMin, _ := big.NewInt(0).SetString("1", 10)
+	mintParams := uniswapv3.INonfungiblePositionManagerMintParams{
+		// The address of the token0, first token of the pool.
+		Token0: poolConfig.Token0.Address,
+		// The address of the token1, second token of the pool.
+		Token1: poolConfig.Token1.Address,
+		// The fee associated with the pool.
+		Fee: poolConfig.Fees,
+
+		// The lower end of the tick range for the position.
+		// Here, we provide liquidity across the whole possible range (divisible by tick spacing).
+		TickLower: tickLower,
+		// The higher end of the tick range for the position.
+		TickUpper: tickUpper,
+
+		// The desired amount of token0 to be sent to the pool during the minting operation.
+		Amount0Desired: poolSize,
+		// The desired amount of token1 to be sent to the pool during the minting operation.
+		Amount1Desired: poolSize,
+		// The minimum acceptable amount of token0 to be sent to the pool. This represents the slippage
+		// protection for token0 during the minting. Here, we mint without any slippage protection.
+		// Don't do this in production!
+		Amount0Min: amMin,
+		// The minimum acceptable amount of token1 to be sent to the pool.
+		Amount1Min: amMin,
+
+		// The destination address of the pool fees.
+		Recipient: recipient,
+
+		// The unix time after which the mint will fail, to protect against long-pending transactions
+		// and wild swings in prices.
+		Deadline: dl, // 10 minutes to execute the swap.
+		// Deadline: big.NewInt(1759474606), // in 2 years (2025-10-03)
+	}
+	if err := blockUntilSuccessful(ctx, c, func() (err error) {
+		// Mint tokens.
+		_, err = nftPositionManagerContract.Mint(tops, mintParams)
+		if err != nil {
+			return err
+		}
+
+		// Check that liquidity has been added to the pool.
+		liquidity, err = poolContract.Liquidity(cops)
+		if err != nil {
+			return err
+		}
+		if liquidity.Cmp(big.NewInt(0)) == 0 {
+			return errors.New("pool has no liquidity")
+		}
+		return nil
+	}); err != nil {
+		log.Error().Err(err).Msg("Unable to provide liquidity to the pool")
+		return err
+	}
+	log.Debug().Interface("liquidity", liquidity).Msg("Liquidity provided to the pool")
+	return nil
+}
